repackage the project to be more reusable
