+++ src/app.ts
@@ -2,8 +2,8 @@ import express, { NextFunction, Request, Response, RequestHandler } from "expres
 import errorHandler from "errorhandler";
 import bodyParser from "body-parser";
 
-import * as submissionAPI from "./infra/controllers/submission";
-
+import * as submissionAPI from "./submission/submission-api";
+import * as schemaApi from "./lectern-client/schema-api";
 import * as middleware from "./middleware";
 import multer from "multer";
 import mongoose from "mongoose";
@@ -25,7 +25,9 @@ app.get("/", (req, res) => res.sendFile(__dirname + "/resources/working.gif"));
 app.get("/submission/registration", middleware.wrapAsync(submissionAPI.getRegistrationByProgramId));
 app.post("/submission/registration", upload.single("registrationFile"), middleware.wrapAsync(submissionAPI.createRegistrationWithTsv));
 app.patch("/submission/registration/:id", middleware.wrapAsync(submissionAPI.commitRegistration));
-app.get("/submission/schema/", middleware.wrapAsync(submissionAPI.getSchema));
+
+/** Schema API */
+app.get("/submission/schema/", middleware.wrapAsync(schemaApi.getSchema));
 
 // this has to be defined after all routes for it to work for these paths.
 app.use(middleware.errorHandler);
